requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31
google-generativeai>=0.5
groq>=0.9
